demos/requirements.txt
rich
kafka-python
clickhouse_connect
glassgen
glassflow-clickhouse-etl
